Move Location module to Game
