+++ bench/Benchmark.hs
@@ -20,7 +20,7 @@ import Swarm.Language.Context qualified as Context
 import Swarm.Language.Pipeline (ProcessedTerm)
 import Swarm.Language.Pipeline.QQ (tmQ)
 import Swarm.Language.Syntax (north)
-import Swarm.Util.Location
+import Swarm.Game.Location
 
 -- | The program of a robot that does nothing.
 idleProgram :: ProcessedTerm
