+++ src/Swarm/Game/Entity.hs
@@ -105,9 +105,9 @@ import Data.Text qualified as T
 import Data.Yaml
 import GHC.Generics (Generic)
 import Swarm.Game.Display
+import Swarm.Game.Location
 import Swarm.Language.Capability
 import Swarm.Util (binTuples, dataNotFound, getDataFileNameSafe, plural, reflow, (?))
-import Swarm.Util.Location
 import Swarm.Util.Yaml
 import Text.Read (readMaybe)
 import Witch
