+++ src/Swarm/TUI/Controller.hs
@@ -71,6 +71,7 @@ import Swarm.Game.Achievement.Definitions
 import Swarm.Game.Achievement.Persistence
 import Swarm.Game.CESK (cancel, emptyStore, initMachine)
 import Swarm.Game.Entity hiding (empty)
+import Swarm.Game.Location
 import Swarm.Game.Robot
 import Swarm.Game.ScenarioInfo
 import Swarm.Game.State
@@ -100,7 +101,6 @@ import Swarm.TUI.Model.UI
 import Swarm.TUI.View (generateModal)
 import Swarm.TUI.View.Objective qualified as GR
 import Swarm.Util hiding ((<<.=))
-import Swarm.Util.Location
 import Swarm.Version (NewReleaseFailure (..))
 import System.Clock
 import System.FilePath (splitDirectories)
