+++ swarm.cabal
@@ -93,6 +93,7 @@ library
                       Swarm.Game.Display
                       Swarm.Game.Entity
                       Swarm.Game.Exception
+                      Swarm.Game.Location
                       Swarm.Game.Log
                       Swarm.Game.Recipe
                       Swarm.Game.Robot
@@ -152,7 +153,6 @@ library
                       Swarm.TUI.View.Objective
                       Swarm.TUI.View.Util
                       Swarm.Util
-                      Swarm.Util.Location
                       Swarm.Util.Yaml
                       Swarm.Version
                       Swarm.Web
